Handled cover upload mostly on client side
