+++ apps/threeid/app/routes/api/update-cover.tsx
@@ -1,0 +1,26 @@
+import { ActionFunction, json } from '@remix-run/cloudflare'
+import { getGalaxyClient } from '~/helpers/galaxyClient'
+import { Visibility } from '~/utils/galaxy.server'
+import { requireJWT } from '~/utils/session.server'
+
+export const action: ActionFunction = async ({ request }) => {
+  const jwt = await requireJWT(request)
+
+  const formData = await request.formData()
+  const coverUrl = formData.get('url') as string
+
+  const galaxyClient = await getGalaxyClient()
+  await galaxyClient.updateProfile(
+    {
+      profile: {
+        cover: coverUrl,
+      },
+      visibility: Visibility.Public,
+    },
+    {
+      'KBT-Access-JWT-Assertion': jwt,
+    }
+  )
+
+  return json(coverUrl)
+}
